Removed AbstractValidationTest and moved functions to GraphParser
