+++ core/src/test/java/nl/tudelft/graphalytics/validation/conn/ConnectedComponentsValidationTest.java
@@ -1,6 +1,6 @@
 package nl.tudelft.graphalytics.validation.conn;
 
-import nl.tudelft.graphalytics.validation.AbstractValidationTest;
+import nl.tudelft.graphalytics.validation.io.GraphParser;
 import nl.tudelft.graphalytics.validation.GraphStructure;
 import nl.tudelft.graphalytics.validation.GraphValues;
 import nl.tudelft.graphalytics.validation.io.LongParser;
@@ -12,7 +12,7 @@ import static org.hamcrest.Matchers.*;
 /**
  * @author Tim Hegeman
  */
-public abstract class ConnectedComponentsValidationTest extends AbstractValidationTest {
+public abstract class ConnectedComponentsValidationTest {
 
 	public abstract ConnectedComponentsOutput executeDirectedConnectedComponents(
 			GraphStructure graph) throws Exception;
@@ -25,7 +25,7 @@ public abstract class ConnectedComponentsValidationTest extends AbstractValidati
 		final String inputPath = "/validation-graphs/conn-dir-input";
 		final String outputPath = "/validation-graphs/conn-dir-output";
 
-		GraphStructure inputGraph = parseGraphStructureFromVertexBasedDataset(
+		GraphStructure inputGraph = GraphParser.parseGraphStructureFromVertexBasedDataset(
 				getClass().getResourceAsStream(inputPath), true);
 
 		ConnectedComponentsOutput executionResult = executeDirectedConnectedComponents(inputGraph);
@@ -38,7 +38,7 @@ public abstract class ConnectedComponentsValidationTest extends AbstractValidati
 		final String inputPath = "/validation-graphs/conn-undir-input";
 		final String outputPath = "/validation-graphs/conn-undir-output";
 
-		GraphStructure inputGraph = parseGraphStructureFromVertexBasedDataset(
+		GraphStructure inputGraph = GraphParser.parseGraphStructureFromVertexBasedDataset(
 				getClass().getResourceAsStream(inputPath), false);
 
 		ConnectedComponentsOutput executionResult = executeUndirectedConnectedComponents(inputGraph);
@@ -48,7 +48,7 @@ public abstract class ConnectedComponentsValidationTest extends AbstractValidati
 
 	private void validateConnectedComponents(ConnectedComponentsOutput executionResult, String outputPath)
 			throws Exception {
-		GraphValues<Long> outputGraph = parseGraphValuesFromDataset(
+		GraphValues<Long> outputGraph = GraphParser.parseGraphValuesFromDataset(
 				getClass().getResourceAsStream(outputPath), new LongParser());
 
 		assertThat("result graph has the correct number of vertices",
